backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary>=2.9.0
python-dotenv==1.0.0
openpyxl==3.1.2
xlrd>=2.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0

# Зависимости для автоматизации обработки заявок на перевозку
# Email обработка (встроенные библиотеки Python, но можно добавить дополнительные)
# imaplib и smtplib встроены в Python
# Для Gmail API (опционально):
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# AI зависимости (опционально - выберите один из вариантов)
# Для OpenAI:
# openai>=1.0.0
# Для Google Gemini:
# google-generativeai>=0.3.0
# Для локальной модели Ollama (уже есть requests):
# requests уже установлен выше

